packages/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
pydantic
python-dotenv